Rearranged ElvMediaWallet commands
